Refactor table properties conversion helpers.
